refactor: rename model.parks to model.parklocation
