+++ app/src/main/java/com/android/streetworkapp/model/parklocation/ParkLocation.kt
@@ -1,3 +1,3 @@
-package com.android.streetworkapp.model.parks
+package com.android.streetworkapp.model.parklocation
 
 data class ParkLocation(val lat: Double, val lon: Double, val id: String)
