+++ app/src/main/java/com/android/streetworkapp/model/parklocation/ParkLocationRepository.kt
@@ -1,4 +1,4 @@
-package com.android.streetworkapp.model.parks
+package com.android.streetworkapp.model.parklocation
 
 /** A repository interface for searching nearby park locations. */
 interface ParkLocationRepository {
