+++ app/src/test/java/com/android/streetworkapp/model/parklocation/OverpassParkLocationRepositoryTest.kt
@@ -1,4 +1,4 @@
-package com.android.streetworkapp.model.parks
+package com.android.streetworkapp.model.parklocation
 
 import okhttp3.Call
 import okhttp3.OkHttpClient
